+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.4-alpha
+* Added CommandManager
+* New commands added
+    * /welcome help - Shows list of commands
+    * /welcome update - Checks for update
+    * /welcome version - Shows installed version of plugin
+    * /welcome showcreditsto [who] - Change option to who show credits (Now only for everyone or newcomers)
+* Added option in config to display credits(or whatever you want) to newcomers
+* Added permissions to commands
+* Added PlaceholdersAPI hook - **NOW SUPPORTING PLACEHOLDERS!**
+* Changes made to config file - **Backup and delete it**
+
 ## 0.3-alpha
 * Added onJoinEvent - when player joins first time, show credits
 * Cleared code from junk